m1_ISPY_processing/requirements.txt
pyarrow<0.16
apache-beam<2.18
black
flake8
google-cloud-storage
google-cloud-bigquery
pydicom
tensorflow-transform
tfx
tensorflow
tensorflow-datasets
